Organization/word2markdown/word2markdown/requirements.txt
dify_plugin>=0.2.0,<0.3.0
requests
python-docx
